+++ CHANGELOG.md
@@ -1,14 +1,14 @@
 # Release Notes
 
 ## [Unreleased](https://github.com/ixocreate/application/compare/0.2.0...develop)
+### Fixed
+- Remove StorageSubManager requirement from Publish Command
 
 ## [v0.2.0 (2019-05-03)](https://github.com/ixocreate/application/compare/0.1.0...0.2.0)
-
 ### Changed
 - Merge PublishDefinitionConfig with PublishConfig
 - Remove Filesystem dependency
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/application/compare/master...0.1.0)
-
 ### Changed
 - Consolidate Package